Move ws server into it's own module.
